fix: 타입,쿼리,병합 추가
