--- apps/hongcha/apis/getWatchList.ts
@@ -1,6 +1,0 @@
-import axios from 'axios';
-
-export const getWatchList = async (pageNum: number | string) => {
-  const { data } = await axios.get(`/api/getWatchList?page=${pageNum}`);
-  return data;
-};
