+++ apps/hongcha/constants/api.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = 'https://api.themoviedb.org/3';
