+++ apps/hongcha/pages/api/getWatchList.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { getWatchList } from 'utils/useAxios';
-import { ResponseDataType } from 'types/apiTypes';
+import { ResponseDataType } from 'types/movieTypes';
 import withBFFHandler from './utils/withBFFHandler';
 
 export default withBFFHandler(async (req: NextApiRequest, res: NextApiResponse) => {
