+++ apps/hongcha/queries/useTopRated.ts
@@ -1,14 +1,14 @@
+import MovieInfoClass from 'models/MovieInfoClass';
 import { useQuery } from '@tanstack/react-query';
 import { getTopRatedList } from 'apis/getTopRatedList';
-import { TopRatedQueryType } from 'types/queryTypes';
-import MovieInfoClass from 'models/MovieInfoClass';
+import { FormattedResponseType } from 'types/responseTypes';
 
 export const useTopRated = (pageNum: number) => {
   const {
     isError,
     isLoading,
     data: { results = [], totalPages = 0 } = {},
-  } = useQuery<TopRatedQueryType, Error, { results: MovieInfoClass[]; totalPages: number }>({
+  } = useQuery<FormattedResponseType, Error, { results: MovieInfoClass[]; totalPages: number }>({
     queryKey: ['topRated', pageNum],
     queryFn: () => getTopRatedList(pageNum),
     select: (data) => ({
