--- apps/hongcha/types/queryTypes.ts
@@ -1,7 +1,0 @@
-import { originTopRatedResultType } from 'types/apiTypes';
-
-// TopRatedQueryType originTopRatedResultType 이름 변경
-export interface TopRatedQueryType {
-  results: originTopRatedResultType[];
-  totalPages: number;
-}
