+++ apps/hongcha/types/requestType.ts
@@ -1,0 +1,12 @@
+interface MediaPostRequest {
+  mediaType: string;
+  mediaId: number;
+}
+
+export interface PostFavoriteType extends MediaPostRequest {
+  favorite: boolean;
+}
+
+export interface PostWatchListType extends MediaPostRequest {
+  watchlist: boolean;
+}
